+++ lib/plausible_web/controllers/api/external_controller.ex
@@ -26,7 +26,7 @@ defmodule PlausibleWeb.Api.ExternalController do
             conn
             |> put_resp_header("x-plausible-dropped", "#{Enum.count(dropped)}")
             |> put_status(400)
-            |> json(%{errors: traverse_errors(first_invalid_changeset)})
+            |> json(%{errors: Plausible.ChangesetHelpers.traverse_errors(first_invalid_changeset)})
           else
             conn
             |> put_resp_header("x-plausible-dropped", "#{Enum.count(dropped)}")
@@ -38,7 +38,7 @@ defmodule PlausibleWeb.Api.ExternalController do
       {:error, %Ecto.Changeset{} = changeset} ->
         conn
         |> put_status(400)
-        |> json(%{errors: traverse_errors(changeset)})
+        |> json(%{errors: Plausible.ChangesetHelpers.traverse_errors(changeset)})
     end
   end
 
@@ -100,16 +100,6 @@ defmodule PlausibleWeb.Api.ExternalController do
         {:error, %Ecto.Changeset{} = changeset} -> changeset
         _ -> false
       end
-    end)
-  end
-
-  defp traverse_errors(changeset) do
-    Ecto.Changeset.traverse_errors(changeset, fn {msg, opts} ->
-      Regex.replace(~r"%{(\w+)}", msg, fn _, key ->
-        opts
-        |> Keyword.get(String.to_existing_atom(key), key)
-        |> to_string()
-      end)
     end)
   end
 end